!include/sys/socktypes.h: change socklen_t to u32
OpenGroup:
To forestall portability problems, it is recommended that applications
not use values larger than 2^31 - 1 for the socklen_t type.

TASK: RTOS-1367

+++ include/sys/socktypes.h
@@ -16,9 +16,9 @@
 #ifndef _SYS_SOCKTYPES_H_
 #define _SYS_SOCKTYPES_H_
 
-#include <stddef.h>
+#include <stdint.h>
 
-typedef size_t socklen_t;
+typedef uint32_t socklen_t;
 
 /* NOTE: Define to make mbedtls recognise socklen_t is defined. */
 #define _SOCKLEN_T
